+++ src/routes/productCategory.ts
@@ -1,11 +1,14 @@
 import express from "express";
 import { deleteModel } from "../controllers/baseController";
 import { createProductCategory, getProductCategories } from "../controllers/productCategory";
+import { authenticateUser, checkCompanyOwnership } from "../middlewares/auth";
 import ProductCategoryModel from "../models/productCategory/model";
 import { validateRequest } from "../validators";
 import { createProductCategoryValidator } from "../validators/productCategory";
 
 const router = express.Router();
+
+router.use(authenticateUser());
 
 router.post('/create', validateRequest(createProductCategoryValidator), createProductCategory);
 
@@ -13,6 +16,7 @@ router.get('/', getProductCategories);
 
 router.delete(
   '/:productCategoryId',
+  checkCompanyOwnership(ProductCategoryModel, 'productCategoryId', 'productCategory'),
   deleteModel(ProductCategoryModel, 'productCategoryId')
 );
