+++ src/routes/productPurchasePrice.ts
@@ -1,18 +1,32 @@
 import express from "express";
 import { deleteModel } from "../controllers/baseController";
 import { createProductPurchasePrice, getProductPurchasePrices } from "../controllers/productPurchasePrice";
+import { authenticateUser, checkCompanyOwnership } from "../middlewares/auth";
+import ProductModel from "../models/product/model";
 import ProductPurchasePriceModel from "../models/productPurchasePrice/model";
 import { validateRequest } from "../validators";
 import { createProductPurchasePriceValidator } from "../validators/productPurchasePrice";
 
 const router = express.Router();
 
-router.post('/:productId/create', validateRequest(createProductPurchasePriceValidator), createProductPurchasePrice);
+router.use(authenticateUser());
 
-router.get('/:productId', getProductPurchasePrices);
+router.post(
+  '/:productId/create',
+  validateRequest(createProductPurchasePriceValidator),
+  checkCompanyOwnership(ProductModel, 'productId', 'product'),
+  createProductPurchasePrice
+);
+
+router.get(
+  '/:productId',
+  checkCompanyOwnership(ProductModel, 'productId', 'product'),
+  getProductPurchasePrices
+);
 
 router.delete(
   '/:productPurchasePriceId',
+  checkCompanyOwnership(ProductPurchasePriceModel, 'productPurchasePriceId', 'productPurchasePrice'),
   deleteModel(ProductPurchasePriceModel, 'productPurchasePriceId')
 );
